Split iterator helpers away into their own module
